pyproject.toml
[project]
name = "clamav_large_archive_scanner"
version = "0.1.0"
authors = [
  { name="Dave Zhu", email="yanbzhu@cisco.com" },
]
description = "A workaround for ClamAV to scan archives > 2 GiB"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD 3-Clause",
    "Operating System :: OS Independent",
]
dependencies = [
    "click~=8.1.7",
    "python-magic~=0.4.27",
    "humanize~=4.9.0",
    "cython~=3.0.6",
    "fastlogging~=1.0.0",
    "colorama~=0.4.6",
]

[project.scripts]
archive = "clamav_large_archive_scanner.main:cli"

[project.urls]
Homepage = "https://github.com/cisco-talos/clamav-large-archive-scanner"
Issues = "https://github.com/cisco-talos/clamav-large-archive-scanner/issues"
